Raise coverage 49%→60%: add 92 service tests + fix 4 bugs
New test files (92 tests):
- test_soul_valuation.py (29): all 4 NILE dimensions + full pipeline
- test_incentive_engine.py (16): points table, awards, false positive penalty
- test_oracle_consensus.py (10): submission, voting, consensus, revaluation
- test_collaboration.py (8): detect→patch, detect→exploit, patch→verify pipelines
- test_pattern_library.py (9): vulnerability pattern storage and retrieval
- test_feedback_loop.py (6): agent accuracy metrics, cross-verification
- test_ecosystem_checker.py (14): Jupiter API, known programs, age estimation

Bugs found and fixed during testing:
- BigInteger PK breaks SQLite autoincrement — switch to Integer in
  agent_message and ecosystem_event models
- collaboration.py payload includes raw UUID objects not JSON-serializable
  — wrap contract_id with str() in all pipeline payloads
- JWT default secret not enforced — add startup warning when
  NILE_JWT_SECRET is default in non-development environments
- Add env field to Settings for environment detection

+++ backend/nile/models/agent_message.py
@@ -3,7 +3,7 @@
 import uuid
 from datetime import UTC, datetime
 
-from sqlalchemy import JSON, BigInteger, DateTime, ForeignKey, String, Uuid, func
+from sqlalchemy import JSON, DateTime, ForeignKey, Integer, String, Uuid, func
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from nile.models.base import Base
@@ -12,7 +12,7 @@ from nile.models.base import Base
 class AgentMessage(Base):
     __tablename__ = "agent_messages"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
 
     sender_agent_id: Mapped[uuid.UUID] = mapped_column(
         Uuid, ForeignKey("agents.id"), nullable=False, index=True
