+++ backend/nile/models/ecosystem_event.py
@@ -3,7 +3,7 @@
 import uuid
 from datetime import UTC, datetime
 
-from sqlalchemy import JSON, BigInteger, DateTime, String, Uuid, func
+from sqlalchemy import JSON, DateTime, Integer, String, Uuid, func
 from sqlalchemy.orm import Mapped, mapped_column
 
 from nile.models.base import Base
@@ -12,7 +12,7 @@ from nile.models.base import Base
 class EcosystemEvent(Base):
     __tablename__ = "ecosystem_events"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
 
     event_type: Mapped[str] = mapped_column(String(64), nullable=False, index=True)
     actor_id: Mapped[uuid.UUID | None] = mapped_column(Uuid)
